pyproject.toml
[tool.poetry]
name = "hse-mlops"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
torch = "^2.5.0"
torchvision = "^0.20.1"
pytorch-lightning = "^2.5.0.post0"
pre-commit = "^4.0.1"
omegaconf = "^2.3.0"
hydra-core = "^1.3.2"
dvc = "^3.59.0"
mlflow = "^2.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
